fix test metric random failed in windows, test=develop

+++ python/paddle/incubate/hapi/tests/test_metrics.py
@@ -71,10 +71,10 @@ class TestAccuracyDynamic(unittest.TestCase):
                 label_var = to_variable(label)
                 pred_var = to_variable(pred)
                 state = to_list(acc.add_metric_op(pred_var, label_var))
-                acc.update(* [s.numpy() for s in state])
+                acc.update(*[s.numpy() for s in state])
                 res_m = acc.accumulate()
                 res_f = accuracy(pred, label, self.topk)
-                assert np.all(np.isclose(np.array(res_m), np.array(res_f), rtol=1e-3)), \
+                assert np.all(np.isclose(np.array(res_m, dtype='float64'), np.array(res_f, dtype='float64'), rtol=1e-3)), \
                         "Accuracy precision error: {} != {}".format(res_m, res_f)
                 acc.reset()
                 assert np.sum(acc.total) == 0
@@ -113,7 +113,7 @@ class TestAccuracyStatic(TestAccuracyDynamic):
             acc.update(*state_ret)
             res_m = acc.accumulate()
             res_f = accuracy(pred, label, self.topk)
-            assert np.all(np.isclose(np.array(res_m), np.array(res_f), rtol=1e-3)), \
+            assert np.all(np.isclose(np.array(res_m, dtype='float64'), np.array(res_f, dtype='float64'), rtol=1e-3)), \
                     "Accuracy precision error: {} != {}".format(res_m, res_f)
             acc.reset()
             assert np.sum(acc.total) == 0
